Fix warning with gcc

+++ examples/deflate.c
@@ -132,7 +132,7 @@ static void deflate(nghttp2_hd_deflater *deflater,
   outlen = (size_t)rv;
 
   printf("\nDeflate (%zu byte(s), ratio %.02f):\n\n", outlen,
-         sum == 0 ? 0 : (double)outlen / sum);
+         sum == 0 ? 0 : (double)outlen / (double)sum);
 
   for (i = 0; i < outlen; ++i) {
     if ((i & 0x0fu) == 0) {
